requirements.txt
Django>=4.2,<6.0
djangorestframework
django-cors-headers
djangorestframework-simplejwt
gunicorn
whitenoise[brotli]
psycopg2-binary
dj-database-url
python-dotenv
requests